Refactor MCP analyze input tool
