+++ crates/rulemorph_mcp/src/tools/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod analyze_input;
 pub(crate) mod list_ops;
